Split DAGMap into a new file and add tests
